+++ plugins/woocommerce-admin/client/launch-your-store/index.js
