+++ plugins/woocommerce-admin/client/launch-your-store/launch-your-store-status/style.scss
