+++ plugins/woocommerce-admin/client/launch-your-store/use-launch-your-store.js
